test(openai): isolate gateway audio pipeline proof
