+++ extensions/openai/realtime-quicksilver-gateway-bridge.test.ts
@@ -1,9 +1,5 @@
 import { spawnSync } from "node:child_process";
 import { describe, expect, it, vi } from "vitest";
-import {
-  OPENAI_QUICKSILVER_MAX_PENDING_AUDIO_BYTES,
-  OPENAI_QUICKSILVER_RELAY_FRAME_BYTES,
-} from "./realtime-quicksilver-audio-buffer.js";
 import { OpenAIQuicksilverGatewayBridge } from "./realtime-quicksilver-gateway-bridge.js";
 import {
   OpenAIQuicksilverAudioPeer,
@@ -583,112 +579,6 @@ describe("GPT-Live gateway relay bridge", () => {
       expect(peer.sendAudio).toHaveBeenCalledTimes(2);
     } finally {
       bridge.close();
-    }
-  });
-
-  it("retains the newest five seconds through delayed peer adoption and the RTP pump", async () => {
-    vi.useFakeTimers();
-    let peerCallbacks:
-      | Parameters<typeof OpenAIQuicksilverAudioPeer.create>[0]["callbacks"]
-      | undefined;
-    let resolvePeer: ((peer: OpenAIQuicksilverAudioPeerContract) => void) | undefined;
-    const peerPromise = new Promise<OpenAIQuicksilverAudioPeerContract>((resolve) => {
-      resolvePeer = resolve;
-    });
-    const onError = vi.fn();
-    const bridge = new OpenAIQuicksilverGatewayBridge({
-      providerConfig: {},
-      model: "gpt-live-1-codex",
-      voice: "marin",
-      audioFormat: { encoding: "pcm16", sampleRateHz: 24_000, channels: 1 },
-      onAudio: vi.fn(),
-      onClearAudio: vi.fn(),
-      onError,
-      runAgentConsult: vi.fn(async () => ({ text: "done" })),
-      logger: { debug: vi.fn(), warn: vi.fn() },
-      resolveAuth: vi.fn(async () => ({
-        type: "oauth" as const,
-        token: "oauth-token",
-        accountId: "account-1",
-      })),
-      createPeer: vi.fn((callbacks) => {
-        peerCallbacks = callbacks;
-        return peerPromise;
-      }),
-      fetchImpl: vi.fn(async () => createCallResponse("v=answer\r\n", "rtc_pending_audio")),
-      webSocketFactory: () => new FakeSocket(),
-    });
-    const connection = bridge.connect();
-    const source = Buffer.alloc(256 * OPENAI_QUICKSILVER_RELAY_FRAME_BYTES);
-    for (let frame = 0; frame < 256; frame += 1) {
-      source
-        .subarray(
-          frame * OPENAI_QUICKSILVER_RELAY_FRAME_BYTES,
-          (frame + 1) * OPENAI_QUICKSILVER_RELAY_FRAME_BYTES,
-        )
-        .fill(frame + 1);
-    }
-    for (let offset = 0; offset < source.length; offset += 8_192) {
-      const callerBuffer = Buffer.from(source.subarray(offset, offset + 8_192));
-      bridge.sendAudio(callerBuffer);
-      callerBuffer.fill(0xff);
-      await vi.advanceTimersByTimeAsync(171);
-    }
-    if (!peerCallbacks) {
-      throw new Error("expected the bridge to start peer creation");
-    }
-    const peer = await OpenAIQuicksilverAudioPeer.create({
-      callbacks: peerCallbacks,
-      iceServers: [],
-    });
-    vi.spyOn(peer, "createOffer").mockResolvedValue("v=offer\r\n");
-    vi.spyOn(peer, "applyAnswer").mockResolvedValue();
-    const testPeer = peer as unknown as TestableAudioPeer;
-    const sendRtp = vi
-      .spyOn(testPeer.state.transceiver.sender, "sendRtp")
-      .mockResolvedValue(undefined);
-    const emittedFrames: Buffer[] = [];
-    const takeNextRelayFrame = testPeer.takeNextRelayFrame.bind(testPeer);
-    vi.spyOn(testPeer, "takeNextRelayFrame").mockImplementation(() => {
-      const frame = takeNextRelayFrame();
-      emittedFrames.push(frame);
-      return frame;
-    });
-    const initialSequenceNumber = testPeer.sequenceNumber;
-    const initialTimestamp = testPeer.timestamp;
-    try {
-      resolvePeer?.(peer);
-      await connection;
-
-      expect(testPeer.pendingAudio).toEqual(
-        source.subarray(source.length - OPENAI_QUICKSILVER_MAX_PENDING_AUDIO_BYTES),
-      );
-
-      testPeer.state.peer.connectionStateChange.execute("connected");
-      await vi.advanceTimersByTimeAsync(4_980);
-
-      expect(testPeer.pendingAudio).toHaveLength(0);
-      expect(emittedFrames).toHaveLength(250);
-      expect(emittedFrames.map((frame) => frame[0])).toEqual([
-        ...Array.from({ length: 249 }, (_, index) => index + 7),
-        0,
-      ]);
-      expect(sendRtp).toHaveBeenCalledTimes(250);
-      const packets = sendRtp.mock.calls.map(
-        ([packet]) =>
-          packet as {
-            header: { payloadType: number; sequenceNumber: number; timestamp: number };
-            payload: Buffer;
-          },
-      );
-      expect(packets.every((packet) => packet.header.payloadType === 111)).toBe(true);
-      expect(packets.every((packet) => packet.payload.length > 0)).toBe(true);
-      expect(packets.at(-1)?.header.sequenceNumber).toBe((initialSequenceNumber + 249) & 0xffff);
-      expect(packets.at(-1)?.header.timestamp).toBe((initialTimestamp + 249 * 960) >>> 0);
-      expect(onError).not.toHaveBeenCalled();
-    } finally {
-      bridge.close();
-      vi.useRealTimers();
     }
   });
